backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioodbc>=0.5.0",
    "alembic>=1.18.1",
    "asyncio>=4.0.0",
    "asyncssh>=2.22.0",
    "asyncwhois>=1.1.12",
    "authlib>=1.6.6",
    "bcrypt>=5.0.0",
    "celery[redis]>=5.6.2",
    "cryptography>=46.0.3",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "eventlet>=0.40.4",
    "fastapi[standard]>=0.128.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "paramiko>=4.0.0",
    "passlib[argon2]>=1.7.4",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pymssql>=2.3.11",
    "pyperclip>=1.11.0",
    "python-multipart>=0.0.21",
    "python-wappalyzer>=0.3.1",
    "python-whois>=0.9.6",
    "setuptools>=80.10.2",
    "sqlmodel>=0.0.31",
    "sqlparse>=0.5.5",
    "sshtunnel>=0.4.0",
    "websockets>=16.0",
]
